Added Quick Sort algorithm and some changes in the others files

+++ include/DataGenerator.hpp
@@ -5,9 +5,9 @@
 
 class DataGenerator{
 private:
-    unsigned int vector_size = 0;
+    size_t vector_size = 0;
 public:
-    DataGenerator(unsigned int size);
+    DataGenerator(size_t size);
     int generate_number_to_vector();
     std::vector<int> generate_vector();
 };
